requirements.txt
# Multi-Task Medical Imaging - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing
Pillow>=9.5.0
opencv-python>=4.8.0

# Utilities
tqdm>=4.65.0
jupyter>=1.0.0
notebook>=6.5.0

# Optional but recommended
ipywidgets>=8.0.0
ipykernel>=6.25.0
